Rabbit MQ Ssl possible fix.
Changed RabbitMQEndpointAddress to create the connection to ssl based on
a parameter in the uri.

rabbitmq://guest:guest:ssl@localhost:5671/ssl_queue

If the uri is constructed with ssl after the password of the use it will
create a connection with ssl enabled and servername to the uri's host.

It's a dirty fix, but didnt found a better way to do it as the rabbitmq
enpoints are constructed just using the uri all around the place,
instead of using the host ssl configuration.